Add custom command pinging
Adds the ability to specify a custom command to ping a user.
This can be done by including the phrase "%mention%" in the
command definition. The ping can then be invoked via either an ID,
the username, or a ping. An absence of these, will omit the %mention%
keyword.

A help command was also added.