Final achievement work
Note: This feature was discussed and eventually decided that it was not
desired. However, given that it is in a function state, and I could see
a future where we would want it, I'm going to throw it up on a remote
feature branch. This feature will live there indefinitely.