Restructure level granting code
Restructures all of the rank/level granting code to be more streamlined.
Uses dataclasses and helper functions to clean up a lot of the cruft
that can gotten in over the years.

One functional change is that users are now checked when they enter the
server for missing roles, including new users who should be given
Villager right away.